Move webpack common to webpack
