+++ packages/webpack-common/src/index.js
@@ -1,3 +1,2 @@
 export { default as parseVersion } from './parseVersion'
 export * from './extractText'
-export * from './webpack'
