--- packages/webpack-common/src/webpack.js
@@ -1,98 +1,0 @@
-/*
- * Functions for use in @webpack-blocks/webpack.
- */
-
-import path from 'path'
-
-export {
-  addPlugins,
-  customConfig,
-  performance,
-  resolveAliases,
-  setContext,
-  setDevTool,
-  setOutput,
-  sourceMaps
-}
-
-/**
- * @see https://webpack.github.io/docs/configuration.html#plugins
- */
-function addPlugins (plugins) {
-  return plugins.length > 0
-    ? () => ({ plugins })
-    : () => ({})          // since webpack-merge would otherwise clear the plugins array
-}
-
-function customConfig (wpConfig) {
-  return () => wpConfig
-}
-
-/**
- * @param {object} performanceBudget
- * @param {number} performanceBudget.maxAssetSize
- * @param {number} performanceBudget.maxEntrypointSize
- * @param {string} performanceBudget.hints              'warning' or 'error'
- */
-function performance (performanceBudget) {
-  return () => ({
-    performance: performanceBudget
-  })
-}
-
-/**
- * @see https://webpack.github.io/docs/configuration.html#resolve-alias
- */
-function resolveAliases (aliases) {
-  return () => ({
-    resolve: {
-      alias: aliases
-    }
-  })
-}
-
-/**
- * @see https://webpack.github.io/docs/configuration.html#context
- */
-function setContext (contextPath) {
-  return () => ({
-    context: path.resolve(contextPath)
-  })
-}
-
-/**
- * @see https://webpack.github.io/docs/configuration.html#devtool
- */
-function setDevTool (devtool) {
-  return () => ({ devtool })
-}
-
-/**
- * @see https://webpack.github.io/docs/configuration.html#output
- */
-function setOutput (output) {
-  if (typeof output === 'string') {
-    output = {
-      filename: path.basename(output) || 'bundle.js',
-      path: path.resolve(path.dirname(output) || './build')
-    }
-  }
-
-  return () => ({ output })
-}
-
-/**
- * Just a convenience wrapper to enable sourcemaps in an easier-to-read fashion
- * than `setDevTool()`.
- * @TODO: Only sets the javascript sourcemaps now. Would be nice to make loaders
- *        enable their specific sourcemaps when `sourceMaps()` is used.
- *
- * @param {string} [devtool]
- * @return {Function}
- */
-function sourceMaps (devtool = 'cheap-module-source-map') {
-  return (context) => {
-    context.sourceMaps = true
-    return { devtool }
-  }
-}
